[MODIFY] 유저 찾기 시 User service 기능을 사용하도록 변경(#134)

+++ src/apis/matches/matchings.resolver.ts
@@ -1,8 +1,9 @@
-import { UseGuards } from '@nestjs/common';
+import { ConflictException, UseGuards } from '@nestjs/common';
 import { ElasticsearchService } from '@nestjs/elasticsearch';
 import { Args, Context, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { GqlAuthAccessGuard } from 'src/commons/auth/gql-auth.guard';
 import { IContext, IElasticSearch } from 'src/commons/type/context';
+import { UsersService } from '../users/user.service';
 import { CreateMatchInput } from './dto/createMatch.input';
 import { UpdateMatchInput } from './dto/updateMatch.input';
 import { Match } from './entities/match.entity';
@@ -12,6 +13,7 @@ import { MatchesService } from './matchings.service';
 export class MatchesResolver {
   constructor(
     private readonly matchesService: MatchesService, //
+    private readonly usersService: UsersService,
     private readonly elasticsearchService: ElasticsearchService,
   ) {}
 
@@ -45,47 +47,69 @@ export class MatchesResolver {
     return this.matchesService.findOne(matchId);
   }
 
-  // @UseGuards(GqlAuthAccessGuard)
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Match)
-  createMatch(
+  async createMatch(
     @Args('createMatchInput') createMatchInput: CreateMatchInput,
-    @Args('email') email: string,
-    // @Context() context: IContext,
+    @Context() context: IContext,
   ) {
-    //나중에 user findOne이 생기면 그거 활용하여 user 가져가기
     //create할 때 이미지도 버킷에 저장하기
+
+    //유저 찾기
+    const user = await this.usersService.findOneByEmail({
+      email: context.req.user.email,
+    });
+    if (!user) {
+      throw new ConflictException('해당 유저의 정보를 찾을 수 없습니다.');
+    }
+
     //새로운 매치 생성
     return this.matchesService.create(
-      createMatchInput,
-      email,
-      //context.req.user.email,
+      createMatchInput, //
+      user,
     );
   }
 
-  // @UseGuards(GqlAuthAccessGuard)
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Match)
-  updateMatch(
+  async updateMatch(
     @Args('matchId') matchId: string,
     @Args('updateMatchInput') updateMatchInput: UpdateMatchInput,
-    @Args('email') email: string,
-    // @Context() context: IContext,
+    @Context() context: IContext,
   ) {
-    //나중에 user findOne해서 작성자가 맞는 지 확인하기
     //update할 때 이미지 버킷에 저장하기
+
+    //유저 찾기
+    const user = await this.usersService.findOneByEmail({
+      email: context.req.user.email,
+    });
+    if (!user) {
+      throw new ConflictException('해당 유저의 정보를 찾을 수 없습니다.');
+    }
+
     //해당 매치 수정
-    return this.matchesService.update(matchId, updateMatchInput, email);
+    return this.matchesService.update(matchId, updateMatchInput, user.email);
   }
 
-  // @UseGuards(GqlAuthAccessGuard)
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean)
   async deleteMatch(
     @Args('matchId') matchId: string, //
+    @Context() context: IContext,
   ) {
     //delete할 때 권한이 있는지 확인하기 -> user findOne 사용해서 가져오기
     //delete할 때 이미지 버킷에 저장하기
 
+    //유저 찾기
+    const user = await this.usersService.findOneByEmail({
+      email: context.req.user.email,
+    });
+    if (!user) {
+      throw new ConflictException('해당 유저의 정보를 찾을 수 없습니다.');
+    }
+
     //매치 데이터 삭제
-    const result = await this.matchesService.delete(matchId);
+    const result = await this.matchesService.delete(matchId, user.email);
 
     //ElasticSearch 데이터 삭제
     await this.elasticsearchService.deleteByQuery({
